Cleanedup and moved the torSessionManager class out of baitConnector.py
